Update code for fsd-ecomm

+++ fsd-demo/be/src/controllers/cartController.js
@@ -5,6 +5,10 @@ export const addToCart = async (req, res) => {
     const { productId, quantity } = req.body
     const product = await Product.findById(productId)
 
+    if (!product) {
+        return res.status(404).json({ message: "Product not found" })
+    }
+
     let cart = await Cart.findOne({ user: req.user._id })
 
     if (!cart) {
@@ -16,7 +20,7 @@ export const addToCart = async (req, res) => {
     )
 
     if (itemIndex > -1) {
-        cart.items[itemIndex].quatity += quantity
+        cart.items[itemIndex].quantity += quantity
     }
     else {
         cart.items.push({
@@ -39,6 +43,11 @@ export const addToCart = async (req, res) => {
 
 export const removeFromCart = async (req, res) => {
     const cart = await Cart.findOne({ user: req.user._id })
+
+    if (!cart) {
+        return res.status(404).json({ message: "Cart not found" })
+    }
+
     await Cart.findByIdAndDelete(cart._id)
 
     res.status(204).json({
